Change user_currency to mandatory and renaming
Change user_currency to mandatory value during setting. Rename
storage.py to db.py within storage module for convenience
